"Refactored authentication and note management API endpoints, updated models, and removed unnecessary files"

+++ server/api/delete.js
@@ -4,8 +4,22 @@ const notesModel = require("../models/notes");
 
 // Route to delete note by ID
 router.delete("/api/delete", async (req, res) => {
-  const { id } = req.body;
+  const { userId, id } = req.body;
+  console.log(req.body);
   try {
+    const note = await notesModel.findOne({ _id: id });
+
+    if (!note) {
+      return res
+        .status(404)
+        .json({ success: false, message: "Note not found" });
+    }
+
+    if (note.userId.toString() !== userId) {
+      return res
+        .status(403)
+        .json({ success: false, message: "Unauthorized to delete this note" });
+    }
     const deleteNote = await notesModel.deleteOne({ _id: id });
     res.status(200).json({ success: true, data: deleteNote });
   } catch (err) {
